+++ apps/stygian/lib/stygian/combat.ex
@@ -7,110 +7,6 @@ defmodule Stygian.Combat do
 
   alias Stygian.Repo
 
-  alias Stygian.Combat.WeaponType
-
-  @doc """
-  Returns the list of weapon_types.
-
-  ## Examples
-
-      iex> list_weapon_types()
-      [%WeaponType{}, ...]
-
-  """
-  def list_weapon_types do
-    Repo.all(WeaponType)
-  end
-
-  @doc """
-  Gets a single weapon_type.
-
-  Raises `Ecto.NoResultsError` if the Weapon type does not exist.
-
-  ## Examples
-
-      iex> get_weapon_type!(123)
-      %WeaponType{}
-
-      iex> get_weapon_type!(456)
-      ** (Ecto.NoResultsError)
-
-  """
-  def get_weapon_type!(id), do: Repo.get!(WeaponType, id)
-
-  @doc """
-  Gets the weapon type by its name.
-  """
-  @spec get_weapon_type_by_name(String.t()) :: WeaponType.t() | nil
-  def get_weapon_type_by_name(name) do
-    Repo.get_by(WeaponType, name: name)
-  end
-
-  @doc """
-  Creates a weapon_type.
-
-  ## Examples
-
-      iex> create_weapon_type(%{field: value})
-      {:ok, %WeaponType{}}
-
-      iex> create_weapon_type(%{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def create_weapon_type(attrs \\ %{}) do
-    %WeaponType{}
-    |> WeaponType.changeset(attrs)
-    |> Repo.insert()
-  end
-
-  @doc """
-  Updates a weapon_type.
-
-  ## Examples
-
-      iex> update_weapon_type(weapon_type, %{field: new_value})
-      {:ok, %WeaponType{}}
-
-      iex> update_weapon_type(weapon_type, %{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def update_weapon_type(%WeaponType{} = weapon_type, attrs) do
-    weapon_type
-    |> WeaponType.changeset(attrs)
-    |> Repo.update()
-  end
-
-  @doc """
-  Deletes a weapon_type.
-
-  ## Examples
-
-      iex> delete_weapon_type(weapon_type)
-      {:ok, %WeaponType{}}
-
-      iex> delete_weapon_type(weapon_type)
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def delete_weapon_type(%WeaponType{} = weapon_type) do
-    Repo.delete(weapon_type)
-  end
-
-  @doc """
-  Returns an `%Ecto.Changeset{}` for tracking weapon_type changes.
-
-  ## Examples
-
-      iex> change_weapon_type(weapon_type)
-      %Ecto.Changeset{data: %WeaponType{}}
-
-  """
-  def change_weapon_type(%WeaponType{} = weapon_type, attrs \\ %{}) do
-    WeaponType.changeset(weapon_type, attrs)
-  end
-
   alias Stygian.Combat.Action
 
   @doc """
